issue #188 bug fix
The default model was set to TheBloke/WizardLM-7B-uncensored-GPTQ which causes issue when running on cpu.

Change the default to TheBloke/vicuna-7B-1.1-HF
When --device_type is cpu or mps, the model_basename will be set to None and will use LlamaForCausalLM. This is a temporary fix. Need a permanent fix for M1/M2.